backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
chromadb==0.4.18
sentence-transformers==2.7.0
huggingface-hub==0.20.2
python-multipart==0.0.6

